Separate HomeScreen from homeScreenStack files
Currently not working, something wrong with exports and imports. I assume its a simple fix or I'm just completely wrong about something in JSX/React Native.
